% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_cells}
\alias{aggregate_cells}
\alias{aggregate_cells,SingleCellExperiment-method}
\title{Aggregate cells}
\usage{
aggregate_cells(.data, slot = "data", assays = NULL,
  aggregation_function = Matrix::rowSums, .by = NULL, .sample = NULL, ...)
}
\arguments{
\item{.data}{A tidySingleCellExperiment object}

\item{.sample}{\lifecycle{soft-deprecated} Use `.by` instead.}

\item{slot}{The slot to which the function is applied}

\item{assays}{The assay to which the function is applied}

\item{aggregation_function}{The method of cell-feature value aggregation}

\item{.by}{Grouping columns (tidyverse-style).}

\item{...}{Used for future extendibility}
}
\value{
A `SummarizedExperiment` object with aggregated assays,
  preserving `rowData` and aggregated `colData`.
}
\description{
Combine cells into groups based on shared variables and aggregate feature counts.
}
\examples{
data(pbmc_small)
pbmc_small_pseudo_bulk <- pbmc_small |>
  aggregate_cells(.by = c(groups, ident), assays = "counts")

}
\references{
Hutchison, W.J., Keyes, T.J., The tidyomics Consortium. et al. The tidyomics ecosystem: enhancing omic data analyses. Nat Methods 21, 1166–1170 (2024). https://doi.org/10.1038/s41592-024-02299-2
}
