% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{append_samples}
\alias{append_samples}
\alias{append_samples.SingleCellExperiment}
\title{Append samples from multiple SingleCellExperiment objects}
\usage{
\method{append_samples}{SingleCellExperiment}(x, ..., .id = NULL)
}
\arguments{
\item{x}{First SingleCellExperiment object to combine}

\item{...}{Additional SingleCellExperiment objects to combine by samples}

\item{.id}{Object identifier (currently not used)}
}
\value{
A combined SingleCellExperiment object
}
\description{
Append samples from multiple SingleCellExperiment objects by column-binding them.
This function is equivalent to `cbind` but provides a tidyverse-like interface.
}
\examples{
data(pbmc_small)
append_samples(pbmc_small, pbmc_small)

}
