% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{relocate}
\alias{relocate}
\title{Change column order}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See *Methods*, below, for
more details.}

\item{...}{<[`tidy-eval`][dplyr_eval]> Variables, or functions or
variables. Use desc() to sort a variable in descending order.}

\item{.before, .after}{<[`tidy-select`][dplyr_tidy_select]> Destination of
columns selected by `...`. Supplying neither will move columns to the
left-hand side; specifying both is an error.}
}
\value{
An object of the same type as `.data`.
}
\description{
Change column order
}
\examples{
`\%>\%` = magrittr::`\%>\%`
pbmc_small \%>\% relocate(groups)
}
\seealso{
Other single table verbs: 
\code{\link{arrange}()},
\code{\link{mutate}()},
\code{\link{rename}()},
\code{\link{slice}()},
\code{\link{summarise}()}
}
\concept{single table verbs}
