% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{join_transcripts}
\alias{join_transcripts}
\title{Add differential transcription information to a tbl using edgeR.}
\usage{
join_transcripts(
  .data,
  transcripts = NULL,
  all = FALSE,
  exclude_zeros = FALSE,
  shape = "long",
  ...
)
}
\arguments{
\item{.data}{A tidy SingleCellExperiment object}

\item{transcripts}{A vector of transcript identifiers to join}

\item{all}{If TRUE return all}

\item{exclude_zeros}{If TRUE exclude zero values}

\item{shape}{Format of the returned table "long" or "wide"}

\item{...}{Parameters to pass to join wide, i.e. assay name to extract transcript abundance from}
}
\value{
A `tbl` containing the information.for the specified transcripts
}
\description{
join_transcripts() extracts and joins information for specific
  transcripts
}
\details{
\lifecycle{experimental}

This function extracts information for specified transcripts and
  returns the information in either long or wide format.
}
\examples{

tidySingleCellExperiment::pbmc_small \%>\%
    
    join_transcripts(transcripts=c("HLA-DRA", "LYZ"))
}
