% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_methods.R
\name{nest}
\alias{nest}
\title{nest}
\arguments{
\item{.data}{A tbl. (See tidyr)}

\item{...}{Name-variable pairs of the form new_col=c(col1, col2, col3) (See tidyr)}

\item{.names_sep}{See ?tidyr::nest}
}
\value{
A tidySingleCellExperiment objector a tibble depending on input
}
\description{
nest
}
\examples{

library(dplyr)
pbmc_small \%>\%

    nest(data=-groups) \%>\%
    unnest(data)
}
