% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{pbmc_small}
\format{
A `SingleCellExperiment` object containing 80 Peripheral Blood 
  Mononuclear Cells (PBMC) from 10x Genomics. Generated by subsampling
  the PBMC dataset of 2,700 single cells.
}
\source{
\url{https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html}
}
\usage{
data(pbmc_small)
}
\value{
`tidySingleCellExperiment`
}
\description{
PBMC single cell RNA-seq data in `SingleCellExperiment` format.
}
\keyword{datasets}
