% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.SingleCellExperiment}
\alias{tidy.SingleCellExperiment}
\title{(DEPRECATED) tidy for `SingleCellExperiment`}
\usage{
\method{tidy}{SingleCellExperiment}(x, ...)
}
\arguments{
\item{x}{A `SingleCellExperiment` object.}

\item{...}{Additional arguments passed to `generics::tidy`. (Unused.)}
}
\value{
A `tidySingleCellExperiment` object. (DEPRECATED - not needed anymore)
}
\description{
(DEPRECATED) tidy for `SingleCellExperiment`
}
\examples{
data(pbmc_small)
pbmc_small

}
\references{
Hutchison, W.J., Keyes, T.J., The tidyomics Consortium. et al. The tidyomics ecosystem: enhancing omic data analyses. Nat Methods 21, 1166–1170 (2024). https://doi.org/10.1038/s41592-024-02299-2
}
