% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy}
\alias{tidy}
\alias{tidy.SingleCellExperiment}
\title{tidy for `SingleCellExperiment`}
\usage{
tidy(object)

\method{tidy}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{A `SingleCellExperiment` object.}
}
\value{
A `tidySingleCellExperiment` object.
}
\description{
tidy for `SingleCellExperiment`
}
\examples{
data(pbmc_small)
pbmc_small

}
