% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy}
\alias{tidy}
\title{tidy for SingleCellExperiment}
\usage{
tidy(object)
}
\arguments{
\item{object}{A SingleCellExperiment object}
}
\value{
A tidySCE object
}
\description{
tidy for SingleCellExperiment
}
\examples{

tidySCE::pbmc_small \%>\% tidy()

}
